% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_concentration_loading.R
\name{custom_concentration_loading}
\alias{custom_concentration_loading}
\title{Open the custom concentration dataset}
\usage{
custom_concentration_loading()
}
\value{
a tibble
}
\description{
`custom_concentration_loading` opens the custom concentration dataset (provided by the Jones/Go Lab, Department of Medicine, Emory University) as a tibble and return it. User can add their own concentration dataset in the same format and load it with this function.
}
