% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.Overlapping.r
\name{find.Overlapping.mzs}
\alias{find.Overlapping.mzs}
\title{find.Overlapping.mzs}
\usage{
find.Overlapping.mzs(dataA, dataB, mz.thresh = 5, time.thresh = NA)
}
\arguments{
\item{dataA}{A feature table. The first column is considered as mz. If the second column
is included (optional), it is considered as retention time.}

\item{dataB}{A feature table. The first column is considered as mz. If the second column
is included (optional), it is considered as retention time.}

\item{mz.thresh}{The m/z threshold for matching features between two data sets.
Default is 5 ppm.}

\item{time.thresh}{The retention time threshold for matching features between two data sets.
Default is NA (retention time is not used for matching). If this is used,
the recommended input is 30 (seconds).}
}
\value{
Matrix of overlapping features with columns:
  index.A, mz.data.A, (optional) time.data.A, index.B, mz.data.B,
  (optional) time.data.B, (optional) time.difference.
}
\description{
This function finds overlapping m/z and/or retention time values between two data sets.
Originally written by Dr. Karan Uppal, Emory University.
Enhanced by James Zhan using data.table.
}
