% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb_concentration_loading.R
\name{hmdb_concentration_loading}
\alias{hmdb_concentration_loading}
\title{Open the HMDB concentration dataset}
\usage{
hmdb_concentration_loading()
}
\value{
a tibble
}
\description{
`hmdb_concentration_loading` opens the HMDB concentration dataset as a tibble and return it.
}
