% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopologue_mass_filter.R
\name{isotopologue_mass_filter}
\alias{isotopologue_mass_filter}
\title{Filter candidate metabolites by most abundant isotopologue mass}
\usage{
isotopologue_mass_filter(
  mz,
  mz_ppm = 10,
  adduct = c("M+H"),
  database = c("metorigindb", "pubchem")
)
}
\arguments{
\item{mz}{The m/z value of the features.}

\item{mz_ppm}{The relative m/z matching threshold for matching features between two data sets. Default is 10 ppm.}

\item{adduct}{The adduct of the m/z. Default is c("M+H"). If multiple adducts are provided, it would be provided as a vector, like c("M+H", "M+Na").}

\item{database}{Database used for metabolite lookup. Choose either
`"metorigindb"` or `"pubchem"`. Default is `"metorigindb"`.}
}
\value{
A data frame of candidate metabolites from the selected database.
}
\description{
`isotopologue_mass_filter` filters candidate metabolites from the selected
database based on the input m/z, adduct, and most abundant isotopologue mass.
The mass-difference estimation window is intentionally wide and estimated
from the m/z and ppm threshold. This serves as a preliminary filter before
precise m/z matching.
}
