% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopologue_mass_match.R
\name{isotopologue_mass_match}
\alias{isotopologue_mass_match}
\title{Match m/z values to candidate metabolites by isotopologue mass}
\usage{
isotopologue_mass_match(
  unknown_feature,
  mz_ppm = 10,
  adduct = c("M+H"),
  database = c("metorigindb", "pubchem")
)
}
\arguments{
\item{unknown_feature}{The mz and time of the unknown features. The first column name has to be  mz and the second column name has to be time.}

\item{mz_ppm}{The m/z ppm threshold for matching the m/z values to the candidate metabolites. Default is 10 ppm.}

\item{adduct}{The adduct of the m/z. Default is c("M+H"). If multiple adducts are provided, it would be provided as a vector, like c("M+H", "M+Na").}

\item{database}{Database used for metabolite lookup. Choose either
`"metorigindb"` or `"pubchem"`. Default is `"metorigindb"`.}
}
\value{
A data frame of mass matched metabolites from the selected database.
}
\description{
`isotopologue_mass_match` first filters the selected metabolite database by
most abundant isotopologue mass with a wide mass window, then performs
precise m/z matching within the specified ppm threshold.
}
