% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_match.R
\name{mass_match}
\alias{mass_match}
\title{Match m/z values to candidate metabolites by monoisotopic mass}
\usage{
mass_match(
  unknown_feature,
  mz_ppm = 10,
  adduct = c("M+H"),
  database = c("metorigindb", "pubchem")
)
}
\arguments{
\item{unknown_feature}{The mz and time of the unknown features. The first column name has to be  mz and the second column name has to be time.}

\item{mz_ppm}{The m/z ppm threshold for matching the m/z values to the candidate metabolites. Default is 10 ppm.}

\item{adduct}{The adduct of the m/z. Default is c("M+H"). If multiple adducts are provided, it would be provided as a vector, like c("M+H", "M+Na").}

\item{database}{Database used for metabolite lookup. Choose either
`"metorigindb"` or `"pubchem"`. Default is `"metorigindb"`.}
}
\value{
A data frame of unknown features matched with molecules from the
  selected database based on the specified m/z ppm threshold and adducts.
}
\description{
`mass_match` first filters the selected metabolite database by monoisotopic
mass with a wide mass window, then performs precise m/z matching within the
specified ppm threshold.
}
