% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_classyfire_loading.R
\name{pubchem_classyfire_loading}
\alias{pubchem_classyfire_loading}
\title{Open the preprocessed PubChem ClassyFire dataset}
\usage{
pubchem_classyfire_loading()
}
\value{
an `arrow::Dataset`
}
\description{
`pubchem_classyfire_loading` opens the PubChem ClassyFire dataset (with ClassyFire classification) preprocessed as an Arrow Dataset, so you can use dplyr verbs without loading the whole file to avoid memory burden.
}
