% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_loading.R
\name{pubchem_loading}
\alias{pubchem_loading}
\title{Open a preprocessed metabolite dataset}
\usage{
pubchem_loading(database = c("metorigindb", "pubchem"))
}
\arguments{
\item{database}{Database used for metabolite lookup. Choose either
`"metorigindb"` or `"pubchem"`. Default is `"metorigindb"`.}
}
\value{
an `arrow::Dataset`
}
\description{
`pubchem_loading` opens the selected metabolite dataset as an Arrow Dataset,
so you can use dplyr verbs without loading the whole file into memory.
}
