% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_mass_search.R
\name{target_mass_search}
\alias{target_mass_search}
\title{Target mass search for the given mass and adduct}
\usage{
target_mass_search(mass, feature_table, adduct = c("M+H"), mz_ppm = 10)
}
\arguments{
\item{mass}{The mass of the targeted chemical.}

\item{feature_table}{The feature table of the untargeted features, assuming the m/z value column is named "mz".}

\item{adduct}{The hypothesized adduct of the chemical.}

\item{mz_ppm}{The m/z ppm threshold for matching the m/z values to the candidate metabolites. Default is 10 ppm.}
}
\value{
A data frame of mass matched (within the specified m/z ppm threshold) features from the feature table.
}
\description{
`target_mass_search` is a function that searches for the given mass and adduct in the feature table.
}
